requirements.txt
diffusers
transformers
accelerate
pydicom
pillow
torch
torchvision
datasets
wandb
pandas
matplotlib
pyyaml
torchmetrics
lpips
clean-fid
